separate out the daily report UI tests
