+++ whatdidUITests/app/PtnViewControllerTest.swift
@@ -261,134 +261,6 @@ class PtnViewControllerTest: UITestBase {
             XCTAssertEqual("whodid", pcombo.stringValue)
         }
     }
-    
-    func testDailyReportPopup() {
-        let entries = EntriesBuilder()
-            .add(project: "project a", task: "task 1", notes: "first thing", minutes: 12)
-            .add(project: "project a", task: "task 2", notes: "sidetrack", minutes: 13)
-            .add(project: "project a", task: "task 1", notes: "back to first", minutes: 5)
-            .add(project: "project b", task: "task 1", notes: "fizz", minutes: 5)
-            .add(project: "project c", task: "task 2", notes: "fuzz", minutes: 10)
-        group("Initalize the data") {
-            let ptn = openPtn()
-            let entriesTextField  = ptn.entriesHook
-            entriesTextField.deleteText(andReplaceWith: FlatEntry.serialize(entries.get()))
-            entriesTextField.typeKey(.enter)
-            clickStatusMenu()
-        }
-        let dailyReport = app.windows[WindowType.dailyEnd.windowTitle]
-        group("Open daily report") {
-            clickStatusMenu(with: .maskAlternate)
-            waitForTransition(of: .dailyEnd, toIsVisible: true)
-        }
-        func verifyProjectsVisibility() {
-            group("Verify projects are visible") {
-                for project in ["project a", "project b", "project c"] {
-                    group(project) {
-                        for (description, e) in HierarchicalEntryLevel(ancestor: dailyReport, scope: "Project", label: project).allElements {
-                            XCTAssertTrue(e.isVisible, "\"\(project)\" \(description) are visible")
-                        }
-                    }
-                }
-            }
-        }
-        verifyProjectsVisibility()
-        group("Verify projects visible after contention with PTN") {
-            group("Close the report and set up contention") {
-                clickStatusMenu() // close daily report
-                let ptn = openPtn()
-                ptn.entriesHook.deleteText(andReplaceWith: FlatEntry.serialize(entries.get(startingAtSecondsSince1970: 86400)) + "\r")
-                setTimeUtc(d: 1, h: 0, m: 0)
-                handleLongSessionPrompt(on: .ptn, .startNewSession)
-                waitForTransition(of: .ptn, toIsVisible: false)
-                waitForTransition(of: .dailyEnd, toIsVisible: true)
-            }
-            verifyProjectsVisibility()
-        }
-        group("Spot check on project a") {
-            let projectA = HierarchicalEntryLevel(ancestor: dailyReport, scope: "Project", label: "project a")
-            let tasksForA = dailyReport.groups["Tasks for \"project a\""]
-            let task1 = HierarchicalEntryLevel(ancestor: tasksForA, scope: "Task", label: "task 1")
-            let task1Details = tasksForA.groups["Details for task 1"]
-            group("Duration label and indicator") {
-                XCTAssertEqual("30m", projectA.durationLabel.stringValue)
-                if let indicatorBarValue = projectA.indicatorBar.value as? Double {
-                    // 30 minutes out of 45 total = 0.6666...
-                    XCTAssertGreaterThan(indicatorBarValue, 0.66)
-                    XCTAssertLessThan(indicatorBarValue, 0.67)
-                }
-            }
-            group("Check tasks for \"project a\"") {
-                XCTAssertFalse(tasksForA.exists)
-                projectA.clickDisclosure(until: tasksForA, .isVisible)
-            }
-            for task in ["task 1", "task 2"] {
-                group("Check \(task)'s visibility") {
-                    for (description, e) in HierarchicalEntryLevel(ancestor: tasksForA, scope: "Task", label: task).allElements {
-                        e.hover()
-                        XCTAssertTrue(e.isVisible, "\(task) \(description)")
-                    }
-                }
-            }
-            group("Spot check on task 1") {
-                group("Duration label and indicator") {
-                    XCTAssertEqual("17m", task1.durationLabel.stringValue)
-                    if let indicatorBarValue = task1.indicatorBar.value as? Double {
-                        // 17 minutes out of 45 total = 0.0.3777...
-                        XCTAssertGreaterThan(indicatorBarValue, 0.37)
-                        XCTAssertLessThan(indicatorBarValue, 0.38)
-                    }
-                }
-                group("Details") {
-                    XCTAssertFalse(task1Details.exists)
-                    task1.clickDisclosure(until: task1Details, .isVisible)
-                    XCTAssertEqual(
-                        [
-                            "1:15am - 1:27am (12m):",
-                            "first thing",
-                            "1:40am - 1:45am (5m):",
-                            "back to first",
-                        ],
-                        task1Details.descendants(matching: .staticText).allElementsBoundByIndex.map({$0.stringValue}))
-                }
-            }
-            group("Task 1 stays expanded if project a folds") {
-                projectA.clickDisclosure(until: task1Details, .doesNotExist)
-                log("Sleeping for a bit to let things stabilize")
-                sleep(2) // Clicking too quickly in a row can break this test
-                projectA.clickDisclosure(until: task1Details, .isVisible)
-            }
-        }
-        group("Projects need to scroll") {
-            group("Set new entries") {
-                clickStatusMenu() // Close the daily report
-                let ptn = openPtn()
-                let manyEntries = EntriesBuilder()
-                for i in 1...25 {
-                    manyEntries.add(project: "project \(i)", task: "only task", notes: "", minutes: Double(i))
-                }
-                ptn.entriesHook.deleteText(andReplaceWith: FlatEntry.serialize(manyEntries.get(startingAtSecondsSince1970: 86400)))
-                ptn.entriesHook.typeKey(.enter)
-                clickStatusMenu()
-            }
-            let project1Header = HierarchicalEntryLevel(ancestor: dailyReport, scope: "Project", label: "project 1").headerLabel
-            group("Open daily report") {
-                clickStatusMenu(with: .maskAlternate)
-                wait(for: "daily report to open", until: {project1Header.exists})
-            }
-            group("Scroll to project 1") {
-                let project25Header = HierarchicalEntryLevel(ancestor: dailyReport, scope: "Project", label: "project 25").headerLabel
-                
-                XCTAssertTrue(project25Header.isVisible)
-                XCTAssertFalse(project1Header.isVisible)
-                
-                project1Header.hover()
-                XCTAssertFalse(project25Header.isVisible)
-                XCTAssertTrue(project1Header.isVisible)
-            }
-        }
-    }
-    
     func testFocus() {
         let ptn = openPtn()
         group("Set up hot key") {
@@ -954,28 +826,6 @@ class PtnViewControllerTest: UITestBase {
     
     func isWindowVisible(_ window: WindowType) -> Bool {
         return app.windows.matching(NSPredicate(format: "title = %@", window.windowTitle)).count > 0
-    }
-    
-    class EntriesBuilder {
-        private var _entries = [(p: String, t: String, n: String, duration: TimeInterval)]()
-        
-        @discardableResult func add(project: String, task: String, notes: String, minutes: Double) -> EntriesBuilder {
-            _entries.append((p: project, t: task, n: notes, duration: minutes * 60.0))
-            return self
-        }
-        
-        func get(startingAtSecondsSince1970 start: Int = 0) -> [FlatEntry] {
-            let totalInterval = _entries.map({$0.duration}).reduce(0, +)
-            var startTime = Date(timeIntervalSince1970: Double(start) - totalInterval)
-            var flatEntries = [FlatEntry]()
-            for e in _entries {
-                let from = startTime
-                let to = startTime.addingTimeInterval(e.duration)
-                flatEntries.append(FlatEntry(from: from, to: to, project: e.p, task: e.t, notes: e.n))
-                startTime = to
-            }
-            return flatEntries
-        }
     }
     
     struct HierarchicalEntryLevel {
